backend/python-scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.18.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.1
webdriver-manager==4.0.1
urllib3==2.0.7
PyPDF2==3.0.1
pdfminer.six==20231228

# Bibliotecas opcionais para processamento avançado de PDFs (descomente para instalar)
# pytesseract==0.3.10  # Para OCR
# pdf2image==1.16.3    # Para converter PDF em imagens para OCR
# Pillow==10.0.0       # Necessário para pdf2image

mysql-connector-python==8.1.0
pytz==2023.3.post1
retrying==1.3.4
cryptography==41.0.4
lxml==4.9.3